Quick and Simple Test to Verify Functionality
Wrote a very quick and easy test for verifying whether or not serial to keyboard input would work. Specifically tested the OBZ 2D barcode scanner and PartsBox as the receiver. The barcode scanner sends the data through serial, which is then received and processed by the script and the keypresses are emulated in the same way a typical barcode scanner works. With that the cheap scanner can now perform the same 'Advanced Data Formatting' that is required for PartsBox.