Start of Console App Interface
In order to make it more usable, the start of a CLI interface has been added. It takes in the port, baud, and key file as arguments and prints some useful information before going into scan mode. In order to be more informative more work needs to be done on the console interface side to give the user more info on what replacements are occurring so they can verify functionality.